Downgrded types.__init__ module from Cython to regulr python
